Add files via upload
Support Vector Machine:

SVM is suspcious. The test score is higher than the training score (should be a little lower).
Moreover, the b is computed using the mean of the biases for each parameter to fit the dimensions of Xw^T-b.
I have to ask someone that know about SVMs.